requirements.txt
PyQt5      5.15.6 
PyQt5-Qt5  5.15.2 
PyQt5-sip  12.10.1
pyzmq      22.3.0 
setuptools 57.0.0 
wheel      0.36.2